requirements.txt
gymnasium[box2d]       
numpy
stable-baselines3
torch
tensorflow
matplotlib
pygame
neat-python     
opencv-python
